committed fixes for the issues reported by analysis tool https://www.pivotaltracker.com/story/show/141761997
note I did not touch the compression code

+++ src/sbml/UnitDefinition.cpp
@@ -998,7 +998,7 @@ UnitDefinition::reorder(UnitDefinition *ud)
   for (n = 0; n < numUnits; n++)
   {
     unit = (Unit *)(units->get(n));
-    int value = unit->getKind();
+    int value = (int)(unit->getKind());
     indexArray[n] = value;
     initialIndexArray[n] = value;
   }
